Final parcial

+++ src/main/java/com/utn/Parcial/Service/PersonService.java
@@ -1,7 +1,10 @@
 package com.utn.Parcial.Service;
 
+import com.utn.Parcial.Exceptions.AlreadyExistException;
 import com.utn.Parcial.Exceptions.NotFoundException;
+import com.utn.Parcial.Model.Manager;
 import com.utn.Parcial.Model.Person;
+import com.utn.Parcial.Model.Player;
 import com.utn.Parcial.Repository.PersonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -9,6 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PersonService {
@@ -21,7 +25,11 @@ public class PersonService {
     }
 
     public void addPerson(Person person){
-        personRepository.save(person);
+        if (this.personRepository.existsById(person.getId())){
+            throw new AlreadyExistException("Person already exist");
+        }else{
+            personRepository.save(person);
+        }
     }
 
     public List<Person> getAll(){
@@ -32,11 +40,31 @@ public class PersonService {
         return personRepository.findById(idPerson).orElseThrow(() -> new NotFoundException("Person not found"));
     }
 
-//    public void addPlayerToPerson(Integer idPerson, Integer idPlayer) {
-//        Person person = this.getPersonById(idPerson);
-//    }
-
     public void deletePerson(Integer idPerson){
-        this.personRepository.deleteById(idPerson);
+        if (this.personRepository.existsById(idPerson)){
+
+            this.personRepository.deleteById(idPerson);
+        }else {
+            throw new NotFoundException("Person doesn't exist");
+        }
+    }
+
+    public void playerByManager(Integer idManager, Integer idPlayer){
+        Person managerOptional = this.getPersonById(idManager);
+        if (managerOptional != null){
+            Manager manager = (Manager) managerOptional;
+            Optional<Player>player = manager.findPlayer(idPlayer);
+            if (player.isPresent()){
+                Player playerOptional = player.get();
+                manager.refreshWeight(playerOptional.getCurrency().getAmount());
+                manager.refreshTotalAmount(playerOptional.getCurrency().getAmount(),
+                                            playerOptional.getCurrency().getCurrency());
+                this.personRepository.save(manager);
+            }else{
+                throw new NotFoundException("Player not found by the manager");
+            }
+        }else {
+            throw new NotFoundException("Manager cannot exists");
+        }
     }
 }
